requirements.txt
python-dotenv==1.0.0
langchain==0.0.189
openai==0.27.7
tiktoken==0.4.0
datasets==2.12.0
apache-beam==2.48.0
ipython==8.14.0